+++ lib/core/database/app_database.dart
@@ -3,6 +3,7 @@ import 'package:path_provider/path_provider.dart';
 import 'package:smart_expense_manager/features/challenges/data/models/weekly_challenge_model.dart';
 import 'package:smart_expense_manager/features/goals/data/models/savings_goal_model.dart';
 import 'package:smart_expense_manager/features/transactions/data/models/category_model.dart';
+import 'package:smart_expense_manager/features/transactions/data/models/merchant_learning_model.dart';
 import 'package:smart_expense_manager/features/transactions/data/models/merchant_rule_model.dart';
 import 'package:smart_expense_manager/features/transactions/data/models/recurring_transaction_model.dart';
 import 'package:smart_expense_manager/features/transactions/data/models/transaction_model.dart';
@@ -17,6 +18,7 @@ class AppDatabase {
         TransactionModelSchema,
         CategoryModelSchema,
         MerchantRuleModelSchema,
+        MerchantLearningModelSchema,
         RecurringTransactionModelSchema,
         SavingsGoalModelSchema,
         WeeklyChallengeModelSchema,